requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
# dlib @ file:///location/dlib-19.24.1-cp311-cp311-win_amd64.whl
face-recognition-models==0.3.0
face_detection==0.2.2
filelock==3.13.1
Flask==3.0.2
fsspec==2024.3.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.2.0
sympy==1.12
torch==2.2.1
torchvision==0.17.1
typing_extensions==4.10.0
Werkzeug==3.0.1